+++ PageRank/VectorMapMatrix.h
@@ -22,14 +22,14 @@ public:
 
     size_t cantColumnas() const;
 
-    void asignar(uint f, uint c, const float value); //Si el valor a asignar puede o no ser 0, usar esta función (y no operator[]).
+    void asignar(uint f, uint c, const double value); //Si el valor a asignar puede o no ser 0, usar esta función (y no operator[]).
 
-    const float at(uint f, uint c) const; //útil si queres leer la posición sin asignar un 0 (operator[] crea el nodo sin importar si no asignas nada).
+    const double at(uint f, uint c) const; //útil si queres leer la posición sin asignar un 0 (operator[] crea el nodo sin importar si no asignas nada).
 
-    float& operator[](pair<uint,uint> p); //Usar solo si se quieren hacer muchas asignaciones distintas de 0.
+    double& operator[](pair<uint,uint> p); //Usar solo si se quieren hacer muchas asignaciones distintas de 0.
     // Cuidado, no usar para asignar ceros, usar asignar en tal caso.
 
-	map<uint, float>& operator[](uint i){ return m[i]; }; //Devuelve una referencia a la i-esima fila.
+	map<uint, double>& operator[](uint i){ return m[i]; }; //Devuelve una referencia a la i-esima fila.
 
     VectorMapMatrix operator+(VectorMapMatrix const &B);
 
@@ -37,20 +37,20 @@ public:
 
     VectorMapMatrix mult(VectorMapMatrix const &B);
 
-    void operator*(float valor);
+    void operator*(double valor);
 
     VectorMapMatrix triangularMatriz();
 
-	pair<vector<float>,short> EG(const VectorMapMatrix& mat, vector<float> b);
+	pair<vector<double>,short> EG(const VectorMapMatrix& mat, vector<double> b);
 
-	pair<vector<float>,short> EGPP(vector<float>);
+	pair<vector<double>,short> EGPP(vector<double>);
 	
 	VectorMapMatrix permutar(unsigned int j, unsigned int i);
 private:
-    vector<map<uint, float> > m; //La matriz va a tener un vector vertical, de arboles rojo negro horizontales.
+    vector<map<uint, double> > m; //La matriz va a tener un vector vertical, de arboles rojo negro horizontales.
     size_t width;
 
-    void resta_de_filas(uint fila_a_modificar, float escalar, uint fila_para_restar);
+    void resta_de_filas(uint fila_a_modificar, double escalar, uint fila_para_restar);
 };
 
 ostream& operator << (ostream &o, VectorMapMatrix &B);
